requirements.txt
# apache-airflow==2.8.1
apache-airflow-providers-http
apache-airflow-providers-postgres
python-dotenv